+++ packages/tools/utils/CHANGELOG.md
@@ -1,6 +1,9 @@
 ## `@svizzle/utils` v0.19.0 (next)
 
 - added `makeTrimmedSplitBy`
+- added `isRegexpEmpty`, `isRegexpNotEmpty`
+- added `makeRegexOf`, `makeSafeRegexOf`
+- added `regexOf`, `safeRegexOf`
 
 ## `@svizzle/utils` v0.18.0